Safari-611.1.10/LayoutTests/fast/css/deferred-parsing/keyframes-rule-expected.txt

This tests deferred parsing of CSS style rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS internals.deferredStyleRulesCount(document.styleSheets[0]) is 0
PASS internals.deferredKeyframesRulesCount(document.styleSheets[0]) is 1
PASS internals.deferredKeyframesRulesCount(document.styleSheets[0]) is 0
PASS successfullyParsed is true

TEST COMPLETE
